💩 User Now has Role Attribution 💩
User has now the feature to add an role when creating a role on BackOffice, of course.

Now you can create an User on backoffice and choose which role do you
want to put on the new user. We are still deciding wheater the admin
needs or not to be there. But soon we update that :).