👔 Added the missing Models 👔
Added all the missing models, now we can start working on Cart and the Payment mecanism.
In a blink of an eye you are going to have payment and express deleviry
in no time.
The favourites is working. Now we just need to make the "Store" Work.

Black Friday is comming earlier this next year

See you Soon C&C Fans