🎨 Introduced Receipt Viewing and Control 🎨
In this patch we created the viewing feature all for you.
Now you can see what you have bought without a Compass, but still with
style.
We created a way to visualize the receipt in a simple way. In the next
patch We will introduce the BIG View of the receipt for now. Just a
litle page to control your expenses.

See you soon C&C Fans