✨ Introduced the printing system to the invoice ✨
Introduced a brand new page for printing and see the invoice in a way just seen in movies.
For you all probably you see this type of invoice everywhere. But for
Cast&Compass we build everything in a professional environment so the
people who use our applications know what they are doing.

Thank you for this journey of knowledge

See you Next Year C&C Fans