🧪 Added a Failing test for User 🧪
Added a Failing test for User and.. so much more
Finally added the Receipt, to the BACKEND.
Added and Table for all employee see the receipts.
Of course... Need the final touch Like the Permissions.
But is all there :)

AND.. Finally.. Added the automated date for all the purchases.
Now is getting ready the Website

See you soon C&C Fans